refactor: Move bot to monorepo structure
This commit refactors the project to a monorepo structure by moving the bot's source code from the `src` directory to `packages/bot`.

The following changes were made:
- The `src` directory has been moved to `packages/bot`.
- The `tests` directory has been moved to `packages/bot/tests`.
- `pyproject.toml` has been updated to reflect the new package structure and entry points.
- The `Dockerfile` has been updated to build the bot from the `packages/bot` directory.
- All internal import statements have been updated from `src.` to `bot.`.
- The `README.md` has been updated to document the new structure.
