+++ packages/bot/llm_services/providers/_common.py
