+++ packages/bot/llm_services/providers/instructions.json
