+++ packages/bot/ryuuko_chatbot.egg-info/entry_points.txt
