+++ packages/ryuuko_bot.egg-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+start = bot.__main__:main
