+++ packages/ryuuko_bot.egg-info/requires.txt
