+++ packages/ryuuko_bot.egg-info/top_level.txt
@@ -1,0 +1,1 @@
+bot
