+++ pyproject.toml
@@ -2,8 +2,12 @@
 requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
 
+[tool.setuptools]
+packages = ["bot"]
+package-dir = {"" = "packages"}
+
 [project]
-name = "ryuuko-chatbot"
+name = "ryuuko-bot"
 version = "1.3.1"
 authors = [
   { name="Zang Vũ", email="zvwgvx@polydevs.uk" },
@@ -39,4 +43,4 @@ test = [
 ]
 
 [project.scripts]
-start = "src.__main__:main"
+start = "bot.__main__:main"
